Introduce PlUploadFormField and PlUploadWidget
PlUploadFormField is a FilePathField that uses the PlUploadWidget.

PlUploadFormField accepts a directory relative to MEDIA_ROOT in path argument.
This argument is passed to the PlUpload javascript widget. This is where
the files will be saved.

PlUploadWidget renders the template that was previously located in
templatetags.py. The javascript PlUploadWidget is fully configurable

At the moment, this technique do not support multi-file uploads.

Add instructions